add authorize to input object
