+++ lib/graphql/schema/input_object.rb
@@ -17,7 +17,9 @@ module GraphQL
           @ruby_style_hash = @arguments.to_kwargs
         end
         # Apply prepares, not great to have it duplicated here.
+        @arguments_by_keyword = {}
         self.class.arguments.each do |name, arg_defn|
+          @arguments_by_keyword[arg_defn.keyword] = arg_defn
           ruby_kwargs_key = arg_defn.keyword
           if @ruby_style_hash.key?(ruby_kwargs_key) && arg_defn.prepare
             @ruby_style_hash[ruby_kwargs_key] = arg_defn.prepare_value(self, @ruby_style_hash[ruby_kwargs_key])
@@ -82,23 +84,19 @@ module GraphQL
         attr_accessor :arguments_class
 
         def argument(name, type, *rest, loads: nil, **kwargs, &block)
-          if loads
-            kwargs[:prepare] ||= Proc.new do |value, context|
-              if value.nil?
-                value
-              elsif value.is_a? Enumerable
-                value.map { |id| context.schema.object_from_id(id, context) }
-              else
-                context.schema.object_from_id(value, context)
-              end
-            end
-          end
-
           argument_defn = super(*argument_with_loads(name, type, *rest, loads: loads, **kwargs, &block))
           # Add a method access
           method_name = argument_defn.keyword
           define_method(method_name) do
-            @ruby_style_hash[method_name]
+            value = @ruby_style_hash[method_name]
+            argument = @arguments_by_keyword[method_name]
+            if loads && argument_defn.type.list?
+              GraphQL::Execution::Lazy.all(value.map { |val| load_application_object(argument, loads, val) })
+            elsif loads
+              load_application_object(argument, loads, value)
+            else
+              value
+            end
           end
         end
